Rename files from PascalCase to snake_case
