+++ backend/hct_mis_api/apps/payment/tests/test_import_export_payment_plan_payment_list.py
@@ -31,13 +31,13 @@ from hct_mis_api.apps.payment.models import (
     ServiceProvider,
 )
 from hct_mis_api.apps.payment.utils import float_to_decimal
-from hct_mis_api.apps.payment.xlsx.XlsxPaymentPlanExportPerFspService import (
+from hct_mis_api.apps.payment.xlsx.xlsx_payment_plan_export_per_fsp_service import (
     XlsxPaymentPlanExportPerFspService,
 )
-from hct_mis_api.apps.payment.xlsx.XlsxPaymentPlanExportService import (
+from hct_mis_api.apps.payment.xlsx.xlsx_payment_plan_export_service import (
     XlsxPaymentPlanExportService,
 )
-from hct_mis_api.apps.payment.xlsx.XlsxPaymentPlanImportService import (
+from hct_mis_api.apps.payment.xlsx.xlsx_payment_plan_import_service import (
     XlsxPaymentPlanImportService,
 )
