+++ backend/hct_mis_api/apps/payment/xlsx/xlsx_payment_plan_export_service.py
@@ -10,7 +10,7 @@ from hct_mis_api.apps.payment.models import (
     Payment,
     PaymentPlan,
 )
-from hct_mis_api.apps.payment.xlsx.BaseXlsxExportService import XlsxExportBaseService
+from hct_mis_api.apps.payment.xlsx.base_xlsx_export_service import XlsxExportBaseService
 from hct_mis_api.apps.payment.xlsx.xlsx_payment_plan_base_service import (
     XlsxPaymentPlanBaseService,
 )
