+++ backend/hct_mis_api/apps/payment/xlsx/xlsx_verification_export_service.py
@@ -15,7 +15,7 @@ from hct_mis_api.apps.payment.models import (
     PaymentVerification,
     XlsxCashPlanPaymentVerificationFile,
 )
-from hct_mis_api.apps.payment.xlsx.BaseXlsxExportService import XlsxExportBaseService
+from hct_mis_api.apps.payment.xlsx.base_xlsx_export_service import XlsxExportBaseService
 
 logger = logging.getLogger(__name__)
